Break up portfolio detail into managable chunks
