+++ frontend/src/components/portfolio/index.js
@@ -1,0 +1,7 @@
+// Portfolio components exports
+export { default as PortfolioSummary } from './PortfolioSummary';
+export { default as PortfolioChart } from './PortfolioChart';
+export { default as FundsTable } from './FundsTable';
+export { default as TransactionsTable } from './TransactionsTable';
+export { default as DividendsTable } from './DividendsTable';
+export { default as PortfolioActions } from './PortfolioActions';
