+++ frontend/src/hooks/portfolio/index.js
@@ -1,0 +1,5 @@
+// Portfolio hooks exports
+export { usePortfolioData } from './usePortfolioData';
+export { useTransactionManagement } from './useTransactionManagement';
+export { useDividendManagement } from './useDividendManagement';
+export { useFundPricing } from './useFundPricing';
