+++ frontend/src/utils/portfolio/index.js
@@ -1,0 +1,4 @@
+// Portfolio utilities exports
+export * from './dateHelpers';
+export * from './portfolioCalculations';
+export * from './transactionValidation';
